wsd: move PrisonerRequestDispatcher to separate files
The COOLWSD.cpp file has become too large. This commit refactors the code by moving the PrisonerRequestDispatcher class into its own dedicated source and header files.

This improves code readability and maintainability without changing the logic.
