+++ Makefile.am
@@ -170,6 +170,7 @@ coolwsd_sources = common/Crypto.cpp \
                   wsd/COOLWSD.cpp \
                   wsd/dumpWsdState.cpp \
                   wsd/ClientRequestDispatcher.cpp \
+                  wsd/PrisonerRequestDispatcher.cpp \
                   wsd/ClientSession.cpp \
                   wsd/DocumentBroker.cpp \
                   wsd/FileServer.cpp \
@@ -393,6 +394,7 @@ wsd_headers = wsd/Admin.hpp \
               wsd/COOLWSD.hpp \
               wsd/COOLWSDServer.hpp \
               wsd/ClientRequestDispatcher.hpp \
+              wsd/PrisonerRequestDispatcher.hpp \
               wsd/ClientSession.hpp \
               wsd/ContentSecurityPolicy.hpp \
               wsd/DocumentBroker.hpp \
